Addressed PR comments IBP-2792
